+++ Cargo.toml
@@ -6,13 +6,13 @@ authors = ["Glenn Watson <gw@intuitionlibrary.com>"]
 [dependencies.freetype]
 git = "https://github.com/servo/rust-freetype"
 
+[dependencies.webrender_traits]
+git = "https://github.com/glennw/webrender_traits"
+
 [dependencies]
 gleam = "*"
-libc = "*"
 euclid = "0.3"
 time = "*"
-serde = "0.6"
-serde_macros = "0.5"
 fnv="*"
 scoped_threadpool = "0.1.6"
 app_units = "0.1"
