+++ src/lib.rs
@@ -1,30 +1,25 @@
-#![feature(step_by, convert, zero_one)]
-#![feature(plugin)]
-#![feature(custom_derive)]
-#![plugin(serde_macros)]
 #![feature(drain)]
 #![feature(hashmap_hasher)]
 #![feature(vec_push_all)]
+#![feature(step_by, convert, zero_one)]
 
-extern crate app_units;
-extern crate euclid;
-extern crate fnv;
-extern crate freetype;
-extern crate gleam;
-extern crate libc;
-extern crate time;
-extern crate serde;
-extern crate scoped_threadpool;
-extern crate simd;
-
-#[cfg(target_os="macos")]
-extern crate core_graphics;
-#[cfg(target_os="macos")]
-extern crate core_text;
-
-pub mod types;
-pub mod renderer;
-pub mod render_api;
+mod aabbtree;
+mod batch;
+mod batch_builder;
+mod clipper;
+mod device;
+mod frame;
+mod freelist;
+mod internal_types;
+mod layer;
+mod node_compiler;
+mod optimizer;
+mod render_backend;
+mod resource_cache;
+mod resource_list;
+mod scene;
+mod texture_cache;
+mod util;
 
 mod platform {
     #[cfg(target_os="macos")]
@@ -42,28 +37,24 @@ mod platform {
     }
 }
 
-mod aabbtree;
-mod batch;
-mod clipper;
-mod device;
-mod freelist;
-mod internal_types;
-mod layer;
-mod optimizer;
-mod render_backend;
-mod resource_cache;
-mod resource_list;
-mod stats;
-mod texture_cache;
-mod util;
+pub mod renderer;
 
-pub use types::{FontKey, ImageKey, StackingLevel, DisplayListID, StackingContext, DisplayListBuilder};
-pub use types::{ColorF, ImageFormat, GradientStop, PipelineId, GlyphInstance, RenderNotifier};
-pub use types::{BorderSide, BorderRadius, BorderStyle, Epoch, BoxShadowClipMode, ClipRegion};
-pub use types::{ScrollLayerId, ScrollPolicy, MixBlendMode, ComplexClipRegion, FilterOp};
-pub use render_api::RenderApi;
+#[cfg(target_os="macos")]
+extern crate core_graphics;
+#[cfg(target_os="macos")]
+extern crate core_text;
+
+extern crate app_units;
+extern crate euclid;
+extern crate fnv;
+extern crate gleam;
+extern crate freetype;
+extern crate scoped_threadpool;
+extern crate simd;
+extern crate time;
+extern crate webrender_traits;
+
 pub use renderer::Renderer;
-
 
 #[doc(hidden)]
 pub mod bench {
@@ -71,4 +62,3 @@ pub mod bench {
     pub use clipper::{clip_polygon, ClipBuffers, Polygon};
     pub use internal_types::WorkVertex;
 }
-
