+++ src/resource_list.rs
@@ -1,12 +1,15 @@
+use aabbtree::AABBTreeNode;
 use app_units::Au;
 use euclid::Size2D;
 use fnv::FnvHasher;
-use internal_types::{BorderRadiusRasterOp, BoxShadowRasterOp};
+use internal_types::{BorderRadiusRasterOp, BoxShadowRasterOp, DrawListItemIndex};
 use internal_types::{Glyph, GlyphKey, RasterItem, TiledImageKey};
+use resource_cache::ResourceCache;
 use std::collections::{HashMap, HashSet};
 use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::hash_state::DefaultState;
-use types::{BorderRadius, FontKey, ImageFormat, ImageKey};
+use webrender_traits::{BorderRadius, BorderStyle, BoxShadowClipMode};
+use webrender_traits::{FontKey, ImageFormat, ImageKey, SpecificDisplayItem};
 
 type RequiredImageSet = HashSet<ImageKey, DefaultState<FnvHasher>>;
 type RequiredGlyphMap = HashMap<FontKey, HashSet<Glyph>, DefaultState<FnvHasher>>;
@@ -168,3 +171,108 @@ impl TiledImageKey {
     }
 }
 
+pub trait BuildRequiredResources {
+    fn build_resource_list(&mut self, resource_cache: &ResourceCache);
+}
+
+impl BuildRequiredResources for AABBTreeNode {
+    fn build_resource_list(&mut self, resource_cache: &ResourceCache) {
+        //let _pf = util::ProfileScope::new("  build_resource_list");
+        let mut resource_list = ResourceList::new();
+
+        for draw_list_index_buffer in &self.draw_lists {
+            let draw_list = resource_cache.get_draw_list(draw_list_index_buffer.draw_list_id);
+
+            for index in &draw_list_index_buffer.indices {
+                let DrawListItemIndex(index) = *index;
+                let display_item = &draw_list.items[index as usize];
+
+                // Handle border radius for complex clipping regions.
+                for complex_clip_region in display_item.clip.complex.iter() {
+                    resource_list.add_radius_raster_for_border_radii(&complex_clip_region.radii);
+                }
+
+                match display_item.item {
+                    SpecificDisplayItem::Image(ref info) => {
+                        resource_list.add_image(info.image_key,
+                                                &display_item.rect.size,
+                                                &info.stretch_size);
+                    }
+                    SpecificDisplayItem::Text(ref info) => {
+                        for glyph in &info.glyphs {
+                            let glyph = Glyph::new(info.size, info.blur_radius, glyph.index);
+                            resource_list.add_glyph(info.font_key, glyph);
+                        }
+                    }
+                    SpecificDisplayItem::Rectangle(..) => {}
+                    SpecificDisplayItem::Gradient(..) => {}
+                    SpecificDisplayItem::BoxShadow(ref info) => {
+                        resource_list.add_radius_raster_for_border_radii(
+                            &BorderRadius::uniform(info.border_radius));
+                        resource_list.add_box_shadow_corner(info.blur_radius,
+                                                            info.border_radius,
+                                                            false);
+                        resource_list.add_box_shadow_edge(info.blur_radius, info.border_radius, false);
+                        if info.clip_mode == BoxShadowClipMode::Inset {
+                            resource_list.add_box_shadow_corner(info.blur_radius,
+                                                                info.border_radius,
+                                                                true);
+                            resource_list.add_box_shadow_edge(info.blur_radius,
+                                                              info.border_radius,
+                                                              true);
+                        }
+                    }
+                    SpecificDisplayItem::Border(ref info) => {
+                        resource_list.add_radius_raster(&info.radius.top_left,
+                                                        &info.top_left_inner_radius(),
+                                                        false,
+                                                        ImageFormat::A8);
+                        resource_list.add_radius_raster(&info.radius.top_right,
+                                                        &info.top_right_inner_radius(),
+                                                        false,
+                                                        ImageFormat::A8);
+                        resource_list.add_radius_raster(&info.radius.bottom_left,
+                                                        &info.bottom_left_inner_radius(),
+                                                        false,
+                                                        ImageFormat::A8);
+                        resource_list.add_radius_raster(&info.radius.bottom_right,
+                                                        &info.bottom_right_inner_radius(),
+                                                        false,
+                                                        ImageFormat::A8);
+
+                        if info.top.style == BorderStyle::Dotted {
+                            resource_list.add_radius_raster(&Size2D::new(info.top.width / 2.0,
+                                                                         info.top.width / 2.0),
+                                                            &Size2D::new(0.0, 0.0),
+                                                            false,
+                                                            ImageFormat::RGBA8);
+                        }
+                        if info.right.style == BorderStyle::Dotted {
+                            resource_list.add_radius_raster(&Size2D::new(info.right.width / 2.0,
+                                                                         info.right.width / 2.0),
+                                                            &Size2D::new(0.0, 0.0),
+                                                            false,
+                                                            ImageFormat::RGBA8);
+                        }
+                        if info.bottom.style == BorderStyle::Dotted {
+                            resource_list.add_radius_raster(&Size2D::new(info.bottom.width / 2.0,
+                                                                         info.bottom.width / 2.0),
+                                                            &Size2D::new(0.0, 0.0),
+                                                            false,
+                                                            ImageFormat::RGBA8);
+                        }
+                        if info.left.style == BorderStyle::Dotted {
+                            resource_list.add_radius_raster(&Size2D::new(info.left.width / 2.0,
+                                                                         info.left.width / 2.0),
+                                                            &Size2D::new(0.0, 0.0),
+                                                            false,
+                                                            ImageFormat::RGBA8);
+                        }
+                    }
+                }
+            }
+        }
+
+        self.resource_list = Some(resource_list);
+    }
+}
